requirements.txt
appdirs==1.4.0
cycler==0.10.0
dask==0.13.0
decorator==4.0.11
Keras==1.2.2
matplotlib==2.0.0
networkx==1.11
numpy==1.12.0
olefile==0.44
opencv-python==3.2.0.6
packaging==16.8
Pillow==4.0.0
protobuf==3.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==0.18.1
six==1.10.0
tensorflow==1.0.0
Theano==0.8.2
toolz==0.8.2
